+++ models/notifications_request/main.go
@@ -2,7 +2,7 @@ package notifications_request
 
 import "github.com/biancarosa/wow-realm-status-notifier/models"
 
-func New(chatID int64, server string) *models.NotificationsRequest {
+func New(chatID uint8, server string) *models.NotificationsRequest {
 	n := new(models.NotificationsRequest)
 	n.ChatID = chatID
 	n.Server = server
